RF: Take `drop`, `uninstall`, `remove` apart into separate modules
This is a precondition for having new-style commands work under PY2
